Add API for http_env_t to own memory
Add http_env_add_header_2 function which applications can use to move
ownership of the memory of the passed header fields and values into
http_env_t. Http_env_t will then use fsfree to deallocate those memory
when the envelope is destroyed.

This allows applications to avoid having to maintain the memory for
headers when they have no other use for them but to pass them into
http_env_t.

Similarly http_env_add_trailer_2 function is added for trailers as the
underlying mechanism for trailers is exactly the same as for headers.